+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [17.20.0] - 2025-03-06
+
+### Changed
+
+- Several improvements in jeap-spring-boot-postgresql-starter:
+  - Schema name is also set to 'data' for JPA. It might require changing the schema name in tests with H2 database, as the default H2 schema name is 'public'.
+  - Removed legacy RDSDataSource implementation. 
+  - Updated AWS advanced jdbc wrapper from 2.5.0 to 2.5.4.
+
 ## [17.19.0] - 2025-03-05
 
 ### Changed